Deal with filth clashes in tests
The passport filth recognises a long integer as being a passport number.
Clearly this will clash with a number of other filths, for example phone
numbers and credit card numbers.

Simplify the tests to allow the outcome of each test to specified
individually, and change expection to deal with clashes.